+++ src/main/java/com/vmo/backendservices/persistance/Repository/EmployeeRepository.java
@@ -1,14 +1,14 @@
-
 package com.vmo.backendservices.persistance.Repository;
 
+import com.vmo.backendservices.persistance.Domain.EmployeeDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.vmo.backendservices.persistance.Domain.EmployeeDetails;;import java.util.List;
+import java.util.List;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<EmployeeDetails, Integer> {
+public interface EmployeeRepository extends JpaRepository<EmployeeDetails, Long> {
 
     @Query("SELECT e.department,count(*) FROM EmployeeDetails e GROUP BY e.department ORDER BY e.department ASC")
     List<Object[]> findDepartmentWithCounts();
