+++ src/main/java/com/vmo/backendservices/persistance/Repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.vmo.backendservices.persistance.Domain.UserLoginInfo;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserLoginInfo, Integer>{
+public interface UserRepository extends JpaRepository<UserLoginInfo, Long>{
 	UserLoginInfo findByUsername(String username);
 }
