+++ src/main/java/com/vmo/backendservices/service/UserDetailService.java
@@ -1,24 +1,22 @@
 package com.vmo.backendservices.service;
 
+import com.vmo.backendservices.persistance.Domain.UserLoginInfo;
+import org.springframework.security.core.userdetails.UserDetails;
+
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.vmo.backendservices.persistance.Domain.UserLoginInfo;
-
 public interface UserDetailService {
-	
-	UserDetails loadUserByUsername(String username);
-	
-	List<UserLoginInfo> findAll();
-	
-	UserLoginInfo save(UserLoginInfo userLoginInfo);
-	
-	Optional<UserLoginInfo> findOne(Integer id);
-	
-	void delete(Integer id);
-	
-	
+
+    UserDetails loadUserByUsername(String username);
+
+    List<UserLoginInfo> findAll();
+
+    UserLoginInfo save(UserLoginInfo userLoginInfo);
+
+    Optional<UserLoginInfo> findOne(Long id);
+
+    void delete(Long id);
+
 
 }
